Move single-use command line equipment export methods out of EquipmentType
